requirements.txt
flask
llama-cpp-python
huggingface_hub